requirements.txt
celery==4.4.0
flower==0.9.3
psycopg2-binary==2.8.4
redis==3.4.1
PyJWT==1.7.1
python-decouple==3.3
Django==2.2.10
django-split-settings==1.0.0
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-allauth==0.41.0
django-rest-auth==0.9.5
oauthlib==3.1.0
requests-oauthlib==1.3.0
django-celery-results==1.2.0
django-filter==2.1.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
whitenoise==5.0.1
django-cors-headers==3.2.1
